fix: check if re-scan required for misconfig reports

+++ pkg/operator/ttl_report.go
@@ -3,10 +3,14 @@ package operator
 import (
 	"context"
 	"fmt"
+	"strconv"
 	"time"
 
+	"github.com/aquasecurity/trivy-operator/pkg/configauditreport"
+	control "github.com/aquasecurity/trivy-operator/pkg/configauditreport/controller"
 	"github.com/aquasecurity/trivy-operator/pkg/ext"
 	"github.com/aquasecurity/trivy-operator/pkg/kube"
+	"github.com/aquasecurity/trivy-operator/pkg/trivyoperator"
 	"github.com/aquasecurity/trivy-operator/pkg/utils"
 
 	"github.com/aquasecurity/trivy-operator/pkg/apis/aquasecurity/v1alpha1"
@@ -24,7 +28,9 @@ import (
 type TTLReportReconciler struct {
 	logr.Logger
 	etc.Config
+	trivyoperator.PluginContext
 	client.Client
+	configauditreport.PluginInMemory
 	ext.Clock
 }
 
@@ -71,20 +77,20 @@ func (r *TTLReportReconciler) DeleteReportIfExpired(ctx context.Context, namespa
 		}
 		return ctrl.Result{}, fmt.Errorf("getting report from cache: %w", err)
 	}
-
 	ttlReportAnnotationStr, ok := reportType.GetAnnotations()[v1alpha1.TTLReportAnnotation]
 	if !ok {
 		log.V(1).Info("Ignoring report without TTL set")
 		return ctrl.Result{}, nil
 	}
-
 	reportTTLTime, err := time.ParseDuration(ttlReportAnnotationStr)
 	if err != nil {
 		return ctrl.Result{}, fmt.Errorf("failed parsing %v with value %v %w", v1alpha1.TTLReportAnnotation, ttlReportAnnotationStr, err)
 	}
 	ttlExpired, durationToTTLExpiration := utils.IsTTLExpired(reportTTLTime, reportType.GetCreationTimestamp().Time, r.Clock)
-	if ttlExpired {
-		log.V(1).Info("Removing report with expired TTL")
+	applicable := r.applicableForDeletion(reportType)
+
+	if ttlExpired && applicable {
+		log.V(1).Info("Removing report with expired TTL or Historical")
 		err := r.Client.Delete(ctx, reportType, &client.DeleteOptions{})
 		if err != nil && !errors.IsNotFound(err) {
 			return ctrl.Result{}, err
@@ -93,5 +99,51 @@ func (r *TTLReportReconciler) DeleteReportIfExpired(ctx context.Context, namespa
 		return ctrl.Result{}, nil
 	}
 	log.V(1).Info("RequeueAfter", "durationToTTLExpiration", durationToTTLExpiration)
+	if ttlExpired {
+		durationToTTLExpiration = reportTTLTime
+	}
 	return ctrl.Result{RequeueAfter: durationToTTLExpiration}, nil
+}
+
+func (r *TTLReportReconciler) applicableForDeletion(report client.Object) bool {
+	reportKind := report.GetObjectKind().GroupVersionKind().Kind
+	if reportKind == "VulnerabilityReport" || reportKind == "ExposedSecretReport" {
+		return true
+	}
+	historicalReportAnnotationStr, ok := report.GetAnnotations()[v1alpha1.HistoricalReportAnnotation]
+	if !ok {
+		historicalReportAnnotationStr = "true" // keep backward compatibale for reports without historical annotation
+	}
+	historical, err := strconv.ParseBool(historicalReportAnnotationStr)
+	if err != nil {
+		return false
+	}
+	if historical {
+		return true
+	}
+	resourceKind, ok := report.GetLabels()[trivyoperator.LabelResourceKind]
+	if !ok {
+		return false
+	}
+	policiesHash, ok := report.GetLabels()[trivyoperator.LabelPluginConfigHash]
+	if !ok {
+		return false
+	}
+	cac, err := r.NewConfigForConfigAudit(r.PluginContext)
+	if err != nil {
+		return false
+	}
+	policies, err := control.Policies(context.Background(), r.Config, r.Client, cac, r.Logger)
+	if err != nil {
+		return false
+	}
+	applicable, err := policies.ExternalPoliciesApplicable(resourceKind)
+	if err != nil {
+		return false
+	}
+	currentPoliciesHash, err := policies.Hash(resourceKind)
+	if err != nil {
+		return false
+	}
+	return applicable && (currentPoliciesHash != policiesHash)
 }
